Move RazorBreakpointSpanEndpoint to VS Contracts
