--- src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Debugging/RazorBreakpointSpanResponse.cs
@@ -1,14 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT license. See License.txt in the project root for license information.
-
-#nullable disable
-
-using OmniSharp.Extensions.LanguageServer.Protocol.Models;
-
-namespace Microsoft.AspNetCore.Razor.LanguageServer.Debugging
-{
-    internal class RazorBreakpointSpanResponse
-    {
-        public Range Range { get; set; }
-    }
-}
