+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/IRazorBreakpointSpanEndpoint.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+using Microsoft.AspNetCore.Razor.LanguageServer.Common;
+using OmniSharp.Extensions.JsonRpc;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
+
+[Parallel, Method(LanguageServerConstants.RazorBreakpointSpanEndpoint)]
+internal interface IRazorBreakpointSpanEndpoint : IJsonRpcRequestHandler<RazorBreakpointSpanParams, RazorBreakpointSpanResponse?>
+{
+}
