+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/EndpointContracts/RazorBreakpointSpanResponse.cs
@@ -1,0 +1,13 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+#nullable disable
+
+using Microsoft.VisualStudio.LanguageServer.Protocol;
+
+namespace Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
+
+internal class RazorBreakpointSpanResponse
+{
+    public Range Range { get; set; }
+}
